+++ packages/react-router/__tests__/router/context-middleware-test.ts
@@ -762,6 +762,10 @@ describe("context/middleware", () => {
           "parent action start",
           "child 1 start - throwing",
           "parent loader start",
+          "child 1 loader start",
+          "child 2 loader start",
+          "child 2 loader end",
+          "child 1 loader end",
           "parent loader end",
         ]);
         expect(router.state.loaderData).toMatchInlineSnapshot(`
@@ -897,6 +901,10 @@ describe("context/middleware", () => {
           "child 2 start",
           "child 2 end - throwing",
           "parent loader start",
+          "child 1 loader start",
+          "child 2 loader start",
+          "child 2 loader end",
+          "child 1 loader end",
           "parent loader end",
         ]);
         expect(router.state.loaderData).toEqual({
@@ -1575,6 +1583,7 @@ describe("context/middleware", () => {
 
     describe("throwing", () => {
       it("throwing from a middleware short circuits immediately (going down - loader)", async () => {
+        let error: unknown;
         let handler = createStaticHandler([
           {
             path: "/",
@@ -1585,6 +1594,12 @@ describe("context/middleware", () => {
             unstable_middleware: [
               async ({ context }, next) => {
                 context.set(parentContext, "PARENT 1");
+                try {
+                  await next();
+                } catch (e) {
+                  error = e;
+                  throw e;
+                }
               },
               async ({ context }, next) => {
                 throw new Error("PARENT 2");
@@ -1624,6 +1639,9 @@ describe("context/middleware", () => {
         expect(staticContext.errors).toEqual({
           parent: "ERROR: PARENT 2",
         });
+
+        // Ensure we don't leak the `middlewareError`structure to userland
+        expect(error).toEqual(new Error("PARENT 2"));
       });
 
       it("throwing from a middleware short circuits immediately (going up - loader)", async () => {